# docs/storoom.api-docs.yaml
openapi: 3.0.1
info:
  title: storoom
  version: v1
servers:
- url: https://localhost:8080/
security:
- Bearer Authentication: []
paths:
...
  /locks/{id}:
    get:
      tags:
      - Locks
      description: Get lock by id.
      operationId: getLock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LockFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
    put:
      tags:
      - Locks
      description: Update lock's info.
      operationId: updateLock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockInfoDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LockFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Locks
      description: Delete lock.
      operationId: deleteLockById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LockFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /locations/{locationId}/owners/{ownerId}:
    put:
      tags:
      - Locations
      description: Add owner to location.
      operationId: addOwnerToLocation
      parameters:
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Locations
      description: Delete owner from location.
      operationId: deleteOwnerFromLocation
      parameters:
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /locations/{id}:
    get:
      tags:
      - Locations
      description: Get location by id.
      operationId: getLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
    put:
      tags:
      - Locations
      description: Update location's info.
      operationId: updateLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Locations
      description: Delete location.
      operationId: deleteLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /companies/{id}:
    put:
      tags:
      - Companies
      description: Update company's info.
      operationId: updateCompany
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompanyReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Companies
      description: Delete company.
      operationId: deleteCompany
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompanyReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
...
  /owners/{ownerId}/locations/{locationId}:
    post:
      tags:
      - Owners
      description: Add location to owner.
      operationId: addLocationToOwner
      parameters:
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Owners
      description: Delete location from owner.
      operationId: deleteLocationFromOwner
      parameters:
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
...
  /locks:
    get:
      tags:
      - Locks
      description: Get all locks.
      operationId: getLocks
      parameters:
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageLockFullDTO'
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Locks
      description: Add lock.
      operationId: createLock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LockFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
  /locations:
    get:
      tags:
      - Locations
      description: Get all locations.
      operationId: getLocations
      parameters:
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationCompactDTO'
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Locations
      description: Add location.
      operationId: addLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
...
  /companies:
    get:
      tags:
      - Companies
      description: Get all companies.
      operationId: getAllCompanies
      parameters:
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyReadDTO'
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Companies
      description: Add company.
      operationId: addCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompanyReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
...
components:
  schemas:
...
    LocationDTO:
      type: object
      properties:
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
        ownerIds:
          type: array
          items:
            type: string
            format: uuid
    LockDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        manufacturer:
          type: string
          format: uuid
...
    LocationCompactDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
    OwnerReadDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        companyId:
          type: string
          format: uuid
        locations:
          type: array
          items:
            $ref: '#/components/schemas/LocationCompactDTO'
...
    ManufacturerDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    LockInfoDTO:
      type: object
      properties:
        name:
          type: string
        manufacturer:
          type: string
          format: uuid
    LockFullDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        manufacturer:
          $ref: '#/components/schemas/ManufacturerDTO'
    LocationReadDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
        owners:
          type: array
          items:
            $ref: '#/components/schemas/OwnerCompactDTO'
    OwnerCompactDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        companyId:
          type: string
          format: uuid
    CompanyDTO:
      type: object
      properties:
        name:
          type: string
    CompanyReadDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageLockFullDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/LockFullDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT
    Basic Authentication:
      type: http
      scheme: basic